backend/requirements.txt
fastapi
uvicorn
easyocr
pdf2image
pytesseract
opencv-python
python-docx
pandas
python-multipart
Pillow
